% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_cor.R
\name{ld_cor}
\alias{ld_cor}
\title{Harmonize LD Matrix with Summary Data}
\usage{
ld_cor(plink2_path, bfile, gwas_sumdata, thread = 80)
}
\arguments{
\item{plink2_path}{The path to the PLINK 2 software.}

\item{bfile}{The binary file of the reference population.}

\item{gwas_sumdata}{The exposure or harmonized summary data.}

\item{thread}{The number of threads for PLINK. Default is 80.}
}
\value{
A matrix of estimated correlations between instrumental variants.
}
\description{
This function extracts instrumental variables from the reference population, aligns them to the same effect allele as the summary dataset, and calculates the correlation between the instrumental variables.
}
\details{
The function ensures that the SNPs in the LD matrix are oriented to the same effect allele as those in the summary dataset.
}
\keyword{IVs}
\keyword{LD}
\keyword{SNP}
\keyword{alignment}
\keyword{correlation}
\keyword{matrix,}
