% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRLOVAR.R
\name{mr_lovar}
\alias{mr_lovar}
\title{Mendelian Randomization Analysis}
\usage{
mr_lovar(
  betaY,
  betaX,
  betaYse,
  betaXse,
  ld_cor,
  ny,
  gwas_p = 0.05,
  gwas_p2 = 5e-08,
  permutn = 0,
  log_file = NULL
)
}
\arguments{
\item{betaY}{beta of the outcome, recommended to be in standardized scale.}

\item{betaX}{beta of the exposure, recommended to be in standardized scale.}

\item{betaYse}{standard error of the outcome, recommended to be in standardized scale.}

\item{betaXse}{standard error of the exposure, recommended to be in standardized scale.}

\item{ld_cor}{LD correlation matrix}

\item{ny}{sample size of outcome GWAS.}

\item{gwas_p}{p-value threshold for the association between variants and the latent phenotype (the direct genetic effect). Default: 5e-2.}

\item{gwas_p2}{p-value threshold for exposure GWAS used in instrument variable selection. Default: 5e-8}

\item{permutn}{number of permutations, Default: 0.}

\item{log_file}{file name for saving the iterations to check convergence. Default: NULL.}
}
\value{
A list that contains
\item{CausEst}{Estimate of causal effect.}
\item{CausEstSE}{Standard error of causal effect estimate.}
\item{CausEstP}{P-value from the z test for the causal effect.}
\item{IVs}{p-values for the direct causal effect of all Instrumental Variants (IV) on the outcome.}
\item{Valid}{Index of valid  instrumental variants.}
\item{sig_v}{the 5th percentile of the permuted p-values.}
\item{corrected_p-value}{corrected p-value based on the permutation distribution.}
}
\description{
This function conducts Mendelian randomization analysis based on the latent phenotype of the outcome, explicitly excluding horizontal pleiotropy variants. It iteratively refines the causal relationship through an Expectation-Maximization (EM) algorithm. MRLOVAR uses GWAS summary statistics and reference panel to account for the dependence between genetic variants as an input to estimate the causal effects of one trait on another. We recommend using summary statistics on a standardized scale.
}
\details{
None.
}
\keyword{Mendelian}
\keyword{randomization.}
